requirement.txt
numpy>=1.21.0
PyQt5>=5.15.0
Pillow>=9.0.0
olefile>=0.46
